BAH-3852 | Refactor. Add subfolder for validation wizard in stock module
