+++ bahmni_stock/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 from . import models
 from . import report
+from . import wizard
