+++ bahmni_stock/models/__init__.py
@@ -5,4 +5,3 @@ from . import stock_move
 from . import stock_move_line
 from . import product
 from . import account_invoice_line
-from . import stock_picking_validate_wizard
