--- bahmni_stock/views/stock_picking_validate_wizard.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="stock_picking_validate_wizard_view" model="ir.ui.view">
-        <field name="name">Validate Stock Picking</field>
-        <field name="model">stock.picking.validate.wizard</field>
-        <field name="arch" type="xml">
-            <form string="Validation">
-                <field name="move_lines">
-                    <tree string="Stock Move Lines">
-                        <field name="product_id"/>
-                        <field name="lot_id" />
-                        <field name="lot_name" string="Lot/Serial Number"/>
-                        <field name="expiration_date"/>
-                        <field name="qty_done" string="Quantity"/>
-                        <field name="product_uom_id"/>
-                        <field name="cost_price"/>
-                        <field name="sale_price"/>
-                        <field name="mrp"/>
-                    </tree>
-                </field>
-                <footer>
-                    <button name="btn_confirm"
-                            string="Confirm"
-                            type="object"
-                            data-hotkey="q"
-                            class="btn-primary"/>
-                    <button string="Discard"
-                            class="btn-secondary"
-                            special="cancel" data-hotkey="z"/>
-                </footer>
-            </form>
-        </field>
-    </record>
-</odoo>
