+++ bahmni_stock/wizard/__init__.py
@@ -1,0 +1,1 @@
+from . import stock_picking_validate_wizard
